refactor: move Table to src/Framework/Database dir
